Moved filter.css to size_t

+++ libsrc/faceflip.cc
@@ -21,9 +21,9 @@ void faceflip(TriMesh *mesh)
 	mesh->tstrips.clear();
 
 	dprintf("Flipping faces... ");
-	int nf = mesh->faces.size();
+	size_t nf = mesh->faces.size();
 #pragma omp parallel for
-	for (int i = 0; i < nf; i++)
+	for (size_t i = 0; i < nf; i++)
 		swap(mesh->faces[i][0], mesh->faces[i][2]);
 	dprintf("Done.\n");
